% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr.plot.R
\name{cgr.plot}
\alias{cgr.plot}
\title{Plot from a CGR object}
\usage{
cgr.plot(data, mode, corners = F, labels = F)
}
\arguments{
\item{data}{CGR object}

\item{mode}{character string
\itemize{
    \item "points": CGR plot
    \item "matrix": FCGR plot
}}

\item{corners}{if true, the corners are added as red dots}

\item{labels}{if true, the symbol associated with the corner is added}
}
\description{
Plot from a CGR object
}
\details{
This function plots the chaos game reprasentation as points or as
frequency matrix representation
}
