% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgr.res.R
\name{cgr.res}
\alias{cgr.res}
\title{Produces a new matrix with new resolution}
\usage{
cgr.res(data, res)
}
\arguments{
\item{data}{CGR object}

\item{res}{new resolution}
}
\description{
Produces a new matrix with new resolution
}
\details{
This function adjust the resolution of the FCGR of a CGR object.

\itemize{
    \item matrix: frequency matrix with new given resolution
    \item x: x-coordinates for the CGR
    \item y: y-coordinates for the CGR
    \item sf: applied scaling factor for the CGR
    \item res: applied resolution to calculate the FCGR
    \item base.seq: chars or letters to build the edges of the CGR
}
}
