% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIV.R
\name{HIV}
\alias{HIV}
\title{Complete HIV 1 genome}
\format{A vector containing the genome of HIV-1 as single characters}
\source{
NC_001802.1 Human immunodeficiency virus 1, complete genome
}
\description{
Complete HIV 1 genome
}
