% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize.R
\name{vectorize}
\alias{vectorize}
\title{Vectorizes the FCGR}
\usage{
vectorize(data)
}
\arguments{
\item{data}{CGR object from cgr.R}
}
\value{
Vector with FCGR-encoding
}
\description{
Vectorizes the FCGR
}
\details{
This function returns the FCGR as vector.
}
